fixed f4560a1 from https://gitee.com/suwenxiang12/graphic_ui/pulls/805 Description: Fallback amplification fontid modification IssueNo: https://gitee.com/openharmony/graphic_ui/issues/I5YRJV Feature or Bugfix: Bugfix Binary Source: No Signed-off-by: suwenxiang <suwenxiang@huawei.com>

+++ frameworks/common/typed_text.h
@@ -40,7 +40,7 @@ public:
     static constexpr uint8_t UTF8_TO_UNICODE_SHIFT3 = 18;
 
     static Point GetTextSize(const char* text,
-                             uint16_t fontId,
+                             uint8_t fontId,
                              uint8_t fontSize,
                              int16_t letterSpace,
                              int16_t lineHeight,
@@ -49,19 +49,19 @@ public:
                              SizeSpan* sizeSpans = nullptr);
 
     static uint32_t GetNextLine(const char* text,
-                                uint16_t fontId,
+                                uint8_t fontId,
                                 uint8_t fontSize,
                                 int16_t letterSpace,
                                 int16_t maxWidth);
 
     static int16_t GetTextWidth(const char* text,
-                                uint16_t fontId,
+                                uint8_t fontId,
                                 uint8_t fontSize,
                                 uint16_t length,
                                 int16_t letterSpace);
 
     static Rect GetArcTextRect(const char* text,
-                                uint16_t fontId,
+                                uint8_t fontId,
                                 uint8_t fontSize,
                                const Point& arcCenter,
                                int16_t letterSpace,
@@ -93,7 +93,7 @@ public:
     static bool IsEmojiModifier(uint32_t codePoint);
     static bool IsEmojiBase(uint32_t codePoint);
 
-    static bool IsColourWord(uint32_t codePoint, uint16_t fontId, uint8_t fontSize);
+    static bool IsColourWord(uint32_t codePoint, uint8_t fontId, uint8_t fontSize);
 
 private:
     static bool GetWrapPoint(const char* text, uint32_t& breakPoint);
