+++ frameworks/font/ui_font_bitmap.cpp
@@ -61,7 +61,7 @@ bool UIFontBitmap::IsVectorFont() const
 }
 
 uint8_t UIFontBitmap::GetShapingFontId(char* text, uint8_t& ttfId, uint32_t& script,
-                                       uint16_t fontId, uint8_t size) const
+                                       uint8_t fontId, uint8_t size) const
 {
 #if defined(ENABLE_MULTI_FONT) && ENABLE_MULTI_FONT
     return UIMultiFontManager::GetInstance()->GetShapingFontId(text, fontId, ttfId, script);
@@ -75,7 +75,7 @@ uint8_t UIFontBitmap::GetShapingFontId(char* text, uint8_t& ttfId, uint32_t& scr
 #endif
 }
 
-uint8_t UIFontBitmap::GetFontWeight(uint16_t fontId)
+uint8_t UIFontBitmap::GetFontWeight(uint8_t fontId)
 {
     UITextLanguageFontParam* fontParam = UIFontBuilder::GetInstance()->GetTextLangFontsTable(fontId);
     if (fontParam == nullptr) {
@@ -111,12 +111,12 @@ int8_t UIFontBitmap::SetFontPath(const char* path, FontType type)
     return RET_VALUE_OK;
 }
 
-uint16_t UIFontBitmap::GetHeight(uint16_t fontId, uint8_t fontSize)
+uint16_t UIFontBitmap::GetHeight(uint8_t fontId, uint8_t fontSize)
 {
     return dynamicFont_.GetFontHeight(fontId);
 }
 
-uint16_t UIFontBitmap::GetFontId(const char* ttfName, uint8_t size) const
+uint8_t UIFontBitmap::GetFontId(const char* ttfName, uint8_t size) const
 {
     if (ttfName == nullptr) {
         return UIFontBuilder::GetInstance()->GetBitmapFontIdMax();
@@ -133,22 +133,22 @@ uint16_t UIFontBitmap::GetFontId(const char* ttfName, uint8_t size) const
     return id;
 }
 
-int16_t UIFontBitmap::GetWidth(uint32_t unicode, uint16_t fontId, uint8_t fontSize)
+int16_t UIFontBitmap::GetWidth(uint32_t unicode, uint8_t fontId, uint8_t fontSize)
 {
     return GetWidthInFontId(unicode, fontId);
 }
 
-uint8_t* UIFontBitmap::GetBitmap(uint32_t unicode, GlyphNode& glyphNode, uint16_t fontId, uint8_t fontSize)
+uint8_t* UIFontBitmap::GetBitmap(uint32_t unicode, GlyphNode& glyphNode, uint8_t fontId, uint8_t fontSize)
 {
     return SearchInFont(unicode, glyphNode, fontId);
 }
 
-bool UIFontBitmap::IsEmojiFont(uint16_t fontId)
+bool UIFontBitmap::IsEmojiFont(uint8_t fontId)
 {
     return false;
 }
 
-int8_t UIFontBitmap::GetFontHeader(FontHeader& fontHeader, uint16_t fontId, uint8_t fontSize)
+int8_t UIFontBitmap::GetFontHeader(FontHeader& fontHeader, uint8_t fontId, uint8_t fontSize)
 {
     const FontHeader* header = dynamicFont_.GetFontHeader(fontId);
     if (header != nullptr) {
@@ -159,7 +159,7 @@ int8_t UIFontBitmap::GetFontHeader(FontHeader& fontHeader, uint16_t fontId, uint
 }
 
 #if defined(ENABLE_MULTI_FONT) && ENABLE_MULTI_FONT
-int8_t UIFontBitmap::GetMultiGlyphNode(uint32_t unicode, GlyphNode& glyphNode, uint16_t fontId)
+int8_t UIFontBitmap::GetMultiGlyphNode(uint32_t unicode, GlyphNode& glyphNode, uint8_t fontId)
 {
     int8_t ret = GetGlyphNode(unicode, glyphNode, fontId);
     if (ret == RET_VALUE_OK) {
@@ -184,7 +184,7 @@ int8_t UIFontBitmap::GetMultiGlyphNode(uint32_t unicode, GlyphNode& glyphNode, u
 }
 #endif
 
-int8_t UIFontBitmap::GetGlyphNode(uint32_t unicode, GlyphNode& glyphNode, uint16_t fontId, uint8_t fontSize)
+int8_t UIFontBitmap::GetGlyphNode(uint32_t unicode, GlyphNode& glyphNode, uint8_t fontId, uint8_t fontSize)
 {
     const GlyphNode* node = dynamicFont_.GetGlyphNode(unicode, fontId);
     if (node != nullptr) {
@@ -223,17 +223,17 @@ int8_t UIFontBitmap::SetCurrentLangId(uint8_t langId)
     return RET_VALUE_OK;
 }
 
-UITextLanguageFontParam* UIFontBitmap::GetFontInfo(uint16_t fontId) const
+UITextLanguageFontParam* UIFontBitmap::GetFontInfo(uint8_t fontId) const
 {
     return UIFontBuilder::GetInstance()->GetTextLangFontsTable(fontId);
 }
 
-int8_t UIFontBitmap::GetDynamicFontBitmap(uint32_t unicode, BufferInfo& bufInfo, uint16_t fontId)
+int8_t UIFontBitmap::GetDynamicFontBitmap(uint32_t unicode, BufferInfo& bufInfo, uint8_t fontId)
 {
     return dynamicFont_.GetBitmap(unicode, bufInfo, fontId);
 }
 
-uint8_t* UIFontBitmap::GetCacheBitmap(uint16_t fontId, uint32_t unicode)
+uint8_t* UIFontBitmap::GetCacheBitmap(uint8_t fontId, uint32_t unicode)
 {
     if (bitmapCache_ != nullptr) {
         return bitmapCache_->GetBitmap(fontId, unicode);
@@ -242,7 +242,7 @@ uint8_t* UIFontBitmap::GetCacheBitmap(uint16_t fontId, uint32_t unicode)
     return nullptr;
 }
 
-BufferInfo UIFontBitmap::GetCacheBuffer(uint16_t fontId, uint32_t unicode, GlyphNode& glyphNode)
+BufferInfo UIFontBitmap::GetCacheBuffer(uint8_t fontId, uint32_t unicode, GlyphNode& glyphNode)
 {
     ColorMode mode = UIFont::GetInstance()->GetColorType(fontId);
     BufferInfo bufInfo{Rect(), 0, nullptr, nullptr, glyphNode.cols, glyphNode.rows, mode, 0};
@@ -267,12 +267,12 @@ void UIFontBitmap::PutCacheSpace(uint8_t* addr)
     GRAPHIC_LOGE("UIFontBitmap::PutCacheSpace invalid bitmapCache");
 }
 
-int16_t UIFontBitmap::GetDynamicFontWidth(uint32_t unicode, uint16_t fontId)
+int16_t UIFontBitmap::GetDynamicFontWidth(uint32_t unicode, uint8_t fontId)
 {
     return dynamicFont_.GetFontWidth(unicode, fontId);
 }
 
-uint8_t* UIFontBitmap::SearchInFont(uint32_t unicode, GlyphNode& glyphNode, uint16_t fontId)
+uint8_t* UIFontBitmap::SearchInFont(uint32_t unicode, GlyphNode& glyphNode, uint8_t fontId)
 {
     GraphicLockGuard guard(lock_);
     if (bitmapCache_ == nullptr) {
@@ -307,7 +307,7 @@ uint8_t* UIFontBitmap::SearchInFont(uint32_t unicode, GlyphNode& glyphNode, uint
     return nullptr;
 }
 
-int16_t UIFontBitmap::GetWidthInFontId(uint32_t unicode, uint16_t fontId)
+int16_t UIFontBitmap::GetWidthInFontId(uint32_t unicode, uint8_t fontId)
 {
     if (!UIFontAdaptor::IsSameTTFId(fontId, unicode)) {
         GRAPHIC_LOGE("UIFontBitmap::GetWidthInFontId fontId and unicode not match");
@@ -324,7 +324,7 @@ void UIFontBitmap::SetFontFileOffset(uint32_t offset)
 uint16_t UIFontBitmap::GetOffsetPosY(const char* text,
                                      uint16_t lineLength,
                                      bool& isEmoijLarge,
-                                     uint16_t fontId,
+                                     uint8_t fontId,
                                      uint8_t fontSize)
 {
     uint32_t i = 0;
@@ -378,7 +378,7 @@ uint16_t UIFontBitmap::GetOffsetPosY(const char* text,
 
 uint16_t UIFontBitmap::GetLineMaxHeight(const char* text,
                                         uint16_t lineLength,
-                                        uint16_t fontId,
+                                        uint8_t fontId,
                                         uint8_t fontSize,
                                         uint16_t& letterIndex,
                                         SizeSpan* sizeSpans)
