+++ interfaces/kits/common/text.h
@@ -121,7 +121,7 @@ struct LineBackgroundColor : public HeapBase {
 struct SizeSpan {
     bool isSizeSpan;
     uint8_t size;
-    uint16_t fontId;
+    uint8_t fontId;
     int16_t height;
 };
 
@@ -205,7 +205,7 @@ public:
      * @since 1.0
      * @version 1.0
      */
-    void SetFontId(uint16_t fontId);
+    void SetFontId(uint8_t fontId);
 
     /**
      * @brief Obtains the font ID.
@@ -518,7 +518,7 @@ protected:
 
     TextStyle* textStyles_;
     char* text_;
-    uint16_t fontId_;
+    uint8_t fontId_;
     uint8_t fontSize_; // Only the vector font library has a valid value.
     Point textSize_;
     bool needRefresh_ : 1;
